refactor: consolidate user authentication handling and remove unused components
