+++ src/styles/globals.css
@@ -144,13 +144,6 @@
     animation: scale-in 0.2s ease-out;
   }
 
-  /* Glass morphism effect */
-  .glass {
-    background: rgba(255, 255, 255, 0.05);
-    backdrop-filter: blur(16px);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-  }
-
   /* Modern gradient text */
   .gradient-text {
     background: linear-gradient(
